requirements.txt
numpy
opencv-python
torch
torchvision
ultralytics
tqdm
pyyaml
